refactor: got rid of ShippingForm
